Update patch set 1
Patch Set 1:

Look at the next commit in the chain, which will be simultaneously submitted. I guess I could add something in this change that is undone in the next. Or squash this into that...

Wanting to submit the next is why I realized I needed this.

Patch-set: 1
Reviewer: Gerrit User 15173 <15173@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>